love redirects to sent + hash_key
