+++ views/web.py
@@ -28,6 +28,7 @@ from logic import TIMESPAN_LAST_WEEK
 from logic import TIMESPAN_THIS_WEEK
 from logic import to_the_future
 from logic import utc_week_limits
+from logic.love_link import create_love_link
 from main import app
 from models import AccessKey
 from models import Alias
@@ -222,19 +223,23 @@ def love():
         return redirect(url_for('home', recipients=recipients_display_str))
 
     try:
-        real_recipients = logic.love.send_loves(recipients, message, secret=secret)
+        real_recipients, hash = logic.love.send_loves(recipients, message, secret=secret)
         # actual recipients may have the sender stripped from the list
         real_display_str = ', '.join(real_recipients)
 
-        flash('{}ove sent to {}!'.format('Secret l' if secret else 'L', real_display_str))
-        return redirect(url_for('home'))
+        if secret:
+            flash('Secret love sent to {}!'.format(real_display_str))
+            return redirect(url_for('home'))
+        else:
+            logging.info(real_recipients)
+            hash_key = create_love_link(','.join(map(str, real_recipients)), message)
+            return redirect(url_for('sent', recipients=recipients_display_str, message=message, hash=hash_key))
+
     except TaintedLove as exc:
         if exc.is_error:
             flash(exc.user_message, 'error')
         else:
             flash(exc.user_message)
-
-        return redirect(url_for('home', recipients=recipients_display_str, message=message))
 
 
 @app.route('/user/autocomplete', methods=['GET'])
